#  Add support for BufferWriter<T> to the JsonWriter where T is IBufferWriter<byte>