--- a/samples/LowAllocationWebServer/Shared/SampleServer.cs
+++ b/samples/LowAllocationWebServer/Shared/SampleServer.cs
@@ -9,7 +9,6 @@ using System.Buffers;
 using System.Threading;
 using System.Text.Http.Parser;
 using Microsoft.Net;
-using System.Buffers.Writer;
 
 namespace LowAllocationWebServer
 {
@@ -63,7 +62,7 @@ namespace LowAllocationWebServer
             response.AppendEoh();
 
             // write response JSON
-            var jsonWriter = new JsonWriter(new BufferWriter<IBufferWriter<byte>>(response), true, prettyPrint: false);
+            JsonWriter<TcpConnectionFormatter> jsonWriter = JsonWriter.Create(response, true, prettyPrint: false);
             jsonWriter.WriteObjectStart();
             jsonWriter.WriteArrayStart("values");
             for (int i = 0; i < 5; i++)
@@ -95,7 +94,7 @@ namespace LowAllocationWebServer
             response.AppendEoh();
 
             // write response JSON
-            var jsonWriter = new JsonWriter(new BufferWriter<IBufferWriter<byte>>(response), true, prettyPrint: false);
+            JsonWriter<TcpConnectionFormatter> jsonWriter = JsonWriter.Create(response, true, prettyPrint: false);
             jsonWriter.WriteObjectStart();
             jsonWriter.WriteArrayStart("values");
             for (int i = 0; i < requestedCount; i++)