Since JsonWriter contains a BufferWriter field now, it has to be a ref struct.

- Maybe `_bufferWriter.Ensure` should return back the span. I tend to call Ensure and then access the buffer right after (_bufferWriter.Buffer).
- Also, calling Flush after each call to Advance is a nuisance.

There is a ~20% regression in performance for the HelloWorld case, across the board (higher constant overhead due to the BufferWriter constructor, roughly 19 ns). However, we get a 15-20% performance improvement for writing a larger json payload. I am not sure why exactly that is. Any ideas?


``` ini

BenchmarkDotNet=v0.10.14.534-nightly, OS=Windows 10.0.14393.0 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=3328125 Hz, Resolution=300.4695 ns, Timer=TSC
.NET Core SDK=2.2.100-preview1-009007
  [Host]     : .NET Core 2.2.0-preview1-26529-01 (CoreCLR 4.6.26525.04, CoreFX 4.6.26525.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.0-preview1-26529-01 (CoreCLR 4.6.26525.04, CoreFX 4.6.26525.01), 64bit RyuJIT


```

**BEFORE:**

|                         Method | IsUTF8Encoded | Formatted |         Mean |       Error |      StdDev |       Median |  Gen 0 | Allocated |
|------------------------------- |-------------- |---------- |-------------:|------------:|------------:|-------------:|-------:|----------:|
|      **WriterSystemTextJsonBasic** |         **False** |     **False** | **11,296.64 ns** | **216.3541 ns** | **212.4887 ns** | **11,262.65 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |     False | 21,245.27 ns | 418.2912 ns | 663.4542 ns | 21,252.39 ns | 0.0916 |     472 B |
| WriterSystemTextJsonHelloWorld |         False |     False |     67.73 ns |   1.3718 ns |   1.2832 ns |     67.62 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |     False |    216.21 ns |   2.9374 ns |   2.4529 ns |    217.23 ns | 0.0739 |     312 B |
|      **WriterSystemTextJsonBasic** |         **False** |      **True** | **13,256.87 ns** | **278.9258 ns** | **481.1333 ns** | **13,191.37 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |      True | 29,857.70 ns | 489.6843 ns | 458.0510 ns | 29,798.08 ns | 0.1221 |     632 B |
| WriterSystemTextJsonHelloWorld |         False |      True |     71.91 ns |   1.4367 ns |   2.0605 ns |     71.70 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |      True |    313.56 ns |   6.2629 ns |  17.5618 ns |    308.08 ns | 0.1121 |     472 B |
|      **WriterSystemTextJsonBasic** |          **True** |     **False** | **11,486.53 ns** | **224.4030 ns** | **275.5871 ns** | **11,413.29 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |     False | 18,858.90 ns | 393.5725 ns | 483.3424 ns | 18,687.17 ns | 0.0916 |     416 B |
| WriterSystemTextJsonHelloWorld |          True |     False |    116.67 ns |   0.7400 ns |   0.6560 ns |    116.68 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |     False |    245.37 ns |   2.3336 ns |   1.9487 ns |    245.17 ns | 0.0606 |     256 B |
|      **WriterSystemTextJsonBasic** |          **True** |      **True** | **12,622.99 ns** |  **17.7916 ns** |  **14.8568 ns** | **12,621.65 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |      True | 27,754.24 ns |  63.0735 ns |  52.6692 ns | 27,741.40 ns | 0.1221 |     576 B |
| WriterSystemTextJsonHelloWorld |          True |      True |    112.05 ns |   0.7270 ns |   0.6444 ns |    111.83 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |      True |    343.16 ns |   6.8418 ns |   7.0260 ns |    344.03 ns | 0.0987 |     416 B |

**AFTER:**

|                         Method | IsUTF8Encoded | Formatted |         Mean |      Error |     StdDev |  Gen 0 | Allocated |
|------------------------------- |-------------- |---------- |-------------:|-----------:|-----------:|-------:|----------:|
|      **WriterSystemTextJsonBasic** |         **False** |     **False** |  **9,538.89 ns** | **186.580 ns** | **242.607 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |     False | 19,977.28 ns | 226.011 ns | 200.353 ns | 0.0916 |     472 B |
| WriterSystemTextJsonHelloWorld |         False |     False |     82.10 ns |   1.630 ns |   1.673 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |     False |    242.20 ns |   2.962 ns |   2.312 ns | 0.0739 |     312 B |
|      **WriterSystemTextJsonBasic** |         **False** |      **True** | **11,179.64 ns** | **185.413 ns** | **164.363 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |      True | 28,603.38 ns | 566.520 ns | 847.940 ns | 0.1221 |     632 B |
| WriterSystemTextJsonHelloWorld |         False |      True |     87.42 ns |   1.712 ns |   2.285 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |      True |    316.48 ns |   5.985 ns |   5.598 ns | 0.1121 |     472 B |
|      **WriterSystemTextJsonBasic** |          **True** |     **False** |  **9,215.93 ns** | **176.737 ns** | **165.320 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |     False | 20,158.62 ns | 402.585 ns | 638.543 ns | 0.0916 |     416 B |
| WriterSystemTextJsonHelloWorld |          True |     False |    142.76 ns |   2.516 ns |   2.354 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |     False |    280.29 ns |   4.935 ns |   6.417 ns | 0.0606 |     256 B |
|      **WriterSystemTextJsonBasic** |          **True** |      **True** | **11,498.84 ns** | **229.579 ns** | **298.518 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |      True | 28,570.19 ns | 551.134 ns | 541.287 ns | 0.1221 |     576 B |
| WriterSystemTextJsonHelloWorld |          True |      True |    142.99 ns |   1.430 ns |   1.268 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |      True |    351.14 ns |   7.372 ns |  13.294 ns | 0.0987 |     416 B |