@JeremyKuhne 2018-06-12
> I am not sure why exactly that is. Any ideas?

Did you try measuring with just changing JsonWriter to a ref struct? No specific reason I think that would make a difference, but I am curious.