@benaadams 2018-06-12
> There is a ~20% regression in performance for the HelloWorld case, across the board (higher constant overhead due to the BufferWriter constructor, roughly 19 ns). However, we get a 15-20% performance improvement for writing a larger json payload.

At a guess the regression is from interface dispatch, and the gain is from buffering in span (reduced virtual dispatch + full advance call costs). Don't think much will be from the constructor.