@benaadams 2018-06-12
Suggested changes hopefully will keep the gains while removing the regressions.

> There is a ~20% regression in performance for the HelloWorld case, across the board (higher constant overhead due to the BufferWriter constructor, roughly 19 ns). However, we get a 15-20% performance improvement for writing a larger json payload.

At a guess the regression is from both boxing the formatters and interface dispatch, and the gain is from buffering in span (reduced costs from less virtual dispatch + full advance calls). Don't think much will be from the BufferWriter constructor.

The suggested changes should remove both the boxing to `IBufferWriter` and the additional interface dispatch costs against those boxes