@benaadams 2018-06-12
> Also, calling Flush after each call to Advance is a nuisance.

The methods shouldn't be calling `Flush`, only `Advance`. 

`Flush` should be done by the caller after a group of operations.