@benaadams 2018-06-12
> Also, calling Flush after each call to Advance is a nuisance.

The methods shouldn't be calling `Flush`, only `Advance`. 

`Flush` should be done by the caller after a group of operations; and before finishing with the `BufferWriter`.

i.e. `Advance` calls "Flush" when it needs to, and an explicit `Flush` is to clear out anything remaining in the Span that hasn't been `Advanced` in the underlying Wrtier (e.g. is still buffered)