@KrzysztofCwalina 2018-06-12
> Since JsonWriter contains a BufferWriter field now, it has to be a ref struct.

The root reason is that now JsonWriter stores/caches the Span. It's done through BufferWriter, but even if you did not use BufferWriter, you would want to cache the span from IBufferWriter, wouldn't you?

On a related note: this PR moves from not caching the span to caching the span and it is only 20% perf improvement? 

Also, writing hello world is what JSON Techempower does and so it would be good to drill into why there is a regression here.

Lastly, I suspect the perf loss is because you now have the overhead of BufferWriter without the benefits of caching the span in cases where the JsonWriter is used to write (access the span) a small number of times. As the number of times you access the span increases, the cost of the buffer writer will be amortized but the cost of individual writes will be smaller.