@ahsonkhan 2018-06-13
> The methods shouldn't be calling `Flush`, only `Advance`. 

> `Flush` should be done by the caller after a group of operations; and before finishing with the `BufferWriter`.

> [EDIT] The current PR calls BufferWriter.Flush in every json write method, this is a perf killer.

> I don't think we can flush on every write. Flush is very expensive when called on Pipelines IBufferWriter. In general, calling flush on every write defeats the purpose of BufferWriter (as Flush will cause Buffer Writer to fish out a new span from Memory)

I agree and it makes sense to me. I will make this change and update the results.

cc @GrabYourPitchforks, since we just talked about this offline as well.