@ahsonkhan 2018-06-13
> The root reason is that now JsonWriter stores/caches the Span. It's done through BufferWriter, but even if you did not use BufferWriter, you would want to cache the span from IBufferWriter, wouldn't you?

> On a related note: this PR moves from not caching the span to caching the span and it is only 20% perf improvement?

We actually end up improving 30%-35%. Here are the updated results with the calls to flush moved to the caller rather than within the JsonWriter APIs (reducing from multiple calls to a single call).

|                         Method | IsUTF8Encoded | Formatted |         Mean |       Error |      StdDev |       Median |  Gen 0 | Allocated |
|------------------------------- |-------------- |---------- |-------------:|------------:|------------:|-------------:|-------:|----------:|
|      **WriterSystemTextJsonBasic** |         **False** |     **False** |  **7,540.44 ns** | **150.3238 ns** | **140.6130 ns** |  **7,587.65 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |     False | 20,027.49 ns | 171.1728 ns | 142.9370 ns | 20,002.59 ns | 0.0916 |     472 B |
| WriterSystemTextJsonHelloWorld |         False |     False |     78.16 ns |   1.5752 ns |   3.1819 ns |     77.67 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |     False |    222.25 ns |   4.2226 ns |   4.1472 ns |    221.30 ns | 0.0741 |     312 B |
|      **WriterSystemTextJsonBasic** |         **False** |      **True** |  **9,551.74 ns** | **155.0085 ns** | **129.4391 ns** |  **9,557.50 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |         False |      True | 29,313.71 ns | 348.3948 ns | 308.8429 ns | 29,253.00 ns | 0.1221 |     632 B |
| WriterSystemTextJsonHelloWorld |         False |      True |     78.19 ns |   1.5312 ns |   1.8228 ns |     78.64 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |         False |      True |    318.94 ns |   1.6915 ns |   1.3206 ns |    318.67 ns | 0.1121 |     472 B |
|      **WriterSystemTextJsonBasic** |          **True** |     **False** |  **7,792.17 ns** | **115.3570 ns** | **102.2609 ns** |  **7,765.99 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |     False | 19,222.77 ns | 444.5917 ns | 371.2542 ns | 19,043.86 ns | 0.0916 |     416 B |
| WriterSystemTextJsonHelloWorld |          True |     False |    125.39 ns |   0.6210 ns |   0.5186 ns |    125.50 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |     False |    261.20 ns |   5.2267 ns |   6.0191 ns |    261.31 ns | 0.0606 |     256 B |
|      **WriterSystemTextJsonBasic** |          **True** |      **True** | **10,282.52 ns** | **202.2167 ns** | **403.8487 ns** | **10,190.40 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |          True |      True | 29,028.33 ns | 574.0739 ns | 589.5316 ns | 29,175.77 ns | 0.1221 |     576 B |
| WriterSystemTextJsonHelloWorld |          True |      True |    143.44 ns |   3.0589 ns |   8.8256 ns |    140.09 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |          True |      True |    359.92 ns |   7.1477 ns |  11.9422 ns |    358.37 ns | 0.0987 |     416 B |